server/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Karthik Macherla", email = "kmacher@seas.upenn.edu"},
]
dependencies = [
    "uvicorn>=0.16.0",
    "requests>=2.26.0",
    "fastapi[all]>=0.58.1",
    "sqlalchemy>=1.4.29",
    "jwt>=1.3.1",
    "configparser>=5.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "google-auth>=2.3.3",
    "bcrypt>=3.2.0",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[project.urls]
homepage = ""

[tool]
[tool.pdm]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
